combine google signup and email signup into one controller and move to users router
