+++ server/controllers/usersController.js
@@ -10,13 +10,76 @@ const jwt = require("jsonwebtoken");
 const {
   sendResetPasswordEmail,
   sendContactUsEmail,
+  sendConfirmationEmail,
 } = require("../utils/nodemailer.config");
+const { verifyGoogleToken } = require("../utils/googleAuth");
 const jwtSecret = process.env.JWT_SECRET;
 
 const getAllUsers = async (req, res, next) => {
   try {
     const users = await User.getAllUsers();
     res.status(200).json(users);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const createUser = async (req, res, next) => {
+  const saltRounds = 12;
+  let userObject;
+  // Google
+  if (req.body.googleCredential) {
+    const verificationRes = await verifyGoogleToken(req.body.googleCredential);
+    if (verificationRes.error) {
+      return res.status(400).json({
+        message: verificationRes.error,
+      });
+    }
+    const profile = verificationRes?.payload;
+    const token = jwt.sign({ email: profile?.email }, jwtSecret);
+
+    userObject = {
+      username:
+        profile?.email.substring(0, profile?.email.indexOf("@")) +
+        "-" +
+        profile.sub.slice(-4),
+      email: profile?.email,
+      firstName: profile?.given_name,
+      lastName: profile?.family_name,
+      photo: profile?.picture,
+      confirmationCode: token,
+      authType: "google",
+      avatar: `/user/avatars/avatar${Math.floor(Math.random() * 12) + 1}.png`,
+    };
+    // Email
+  } else {
+    // TODO: Implement form validation
+    const { username, email, password } = req.body;
+    const token = jwt.sign({ email }, jwtSecret);
+
+    const hashedPassword = await bcrypt.hash(password, saltRounds);
+    userObject = {
+      username,
+      email,
+      hashedPassword,
+      confirmationCode: token,
+      authType: "email",
+      avatar: `/user/avatars/avatar${Math.floor(Math.random() * 12) + 1}.png`,
+    };
+  }
+  const userCheck = await User.getUserByEmail(userObject.email);
+  if (userCheck) {
+    return res.status(401).json({
+      message: "Account already exists. Please login.",
+    });
+  }
+  const user = await User.createUser(userObject);
+  await sendConfirmationEmail(user.username, user.email, user.confirmationCode);
+  res.status(201).json({
+    message:
+      "User was registered successfully. Please check your email to verify.",
+  });
+  try {
   } catch (error) {
     next(error);
   }
@@ -273,6 +336,7 @@ const updatePassword = async (req, res, next) => {
 
 module.exports = {
   getAllUsers,
+  createUser,
   getPublicUser,
   updateUserResetPassword,
   resetUserPassword,
